parking-lot/src/main/java/executors/ExitCommandExecutor.java
package executors;

import model.Command;
import printer.OutputPrinter;
import service.ParkingLotService;

public class ExitCommandExecutor extends CommandExecutor {

    public static final String COMMAND_NAME = "exit";

    public ExitCommandExecutor(ParkingLotService parkingLotService, OutputPrinter outputPrinter) {
        super(parkingLotService, outputPrinter);
    }

    @Override
    public boolean isValid(Command command) {
        return command.getParams().isEmpty();
    }

    @Override
    public void execute(Command command) {
        outputPrinter.end();
    }
}
